# Add an optimized Attributes implementation for instrumenter